rename: list 반환 apiHandler 이름 변경
