+++ app/src/components/data/GetNodeList.vue
@@ -34,7 +34,7 @@
 
 import DataCardList from '@/components/common/DataCardList.vue';
 import type { GetNodeListRequestParam, GetNodeListResponseDto } from '@/lib/api/node/nodeDto';
-import { getNodesApi } from '@/lib/api/node/getNodesApi';
+import { getNodeListApi } from '@/lib/api/node/getNodeListApi';
 
 defineProps<{
   req: GetNodeListRequestParam
@@ -54,7 +54,7 @@ const isEmpty = (item: boolean) => {
 }
 
 const dataHandler = async (req: GetNodeListRequestParam) => {
-  const res = await getNodesApi(req);
+  const res = await getNodeListApi(req);
   if (res) return res.map(({ clusterId, ...rest }) => rest);
 }
