+++ app/src/lib/api/perm/getPermListApi.ts
@@ -2,18 +2,17 @@ import { apiRequest } from "@/lib/global/apiHandler";
 import { ApiError } from "@/lib/global/customError";
 import { useDialogStore } from "@/store/dialog";
 import type { ApiResponse } from "@/lib/global/ApiResponse";
-import type { GetRoleListResponseDto } from "@/lib/api/role/roleDto";
+import type { GetPermListResponseDto } from "@/lib/api/perm/permDto";
 
-
-export async function getRolesApi():Promise<GetRoleListResponseDto[] | null> {
+export async function getPermListApi():Promise<GetPermListResponseDto[] | null> {
 
   const dialog = useDialogStore();
 
   return await apiRequest({
     method: "GET",
-    path: `api/roles`,
+    path: `api/permissions`,
     auth: true
-  }).then(async (res: ApiResponse<GetRoleListResponseDto[]>) => {
+  }).then(async (res: ApiResponse<GetPermListResponseDto[]>) => {
     if (res.data) {
       return res.data;
     }
@@ -21,7 +20,7 @@ export async function getRolesApi():Promise<GetRoleListResponseDto[] | null> {
   }).catch(async (e: ApiError) => {
     console.error(e.res);
     dialog.open({
-      title: '역할 리스트 조회 실패',
+      title: '권한 리스트 조회 실패',
       message: e.res.message,
       type: 'mainframe'
     });
