+++ app/src/lib/api/role/getRoleListApi.ts
@@ -2,23 +2,26 @@ import { apiRequest } from "@/lib/global/apiHandler";
 import { ApiError } from "@/lib/global/customError";
 import { useDialogStore } from "@/store/dialog";
 import type { ApiResponse } from "@/lib/global/ApiResponse";
-import type { GetNodeListRequestParam, GetNodeListResponseDto } from "@/lib/api/node/nodeDto";
+import type { GetRoleListResponseDto } from "@/lib/api/role/roleDto";
 
-export async function getNodesApi(reqDto: GetNodeListRequestParam):Promise<GetNodeListResponseDto[] | null> {
+
+export async function getRoleListApi():Promise<GetRoleListResponseDto[] | null> {
 
   const dialog = useDialogStore();
 
   return await apiRequest({
     method: "GET",
-    path: `api/clusters/${reqDto.clusterId}/nodes`,
+    path: `api/roles`,
     auth: true
-  }).then(async (res: ApiResponse<GetNodeListResponseDto[]>) => {
-    if (!res.data) return null;
-    return res.data;
+  }).then(async (res: ApiResponse<GetRoleListResponseDto[]>) => {
+    if (res.data) {
+      return res.data;
+    }
+    return null;
   }).catch(async (e: ApiError) => {
     console.error(e.res);
     dialog.open({
-      title: '노드 리스트 조회 실패',
+      title: '역할 리스트 조회 실패',
       message: e.res.message,
       type: 'mainframe'
     });
